refactor: 결제 클라이언트 인터페이스화
